@@ -1,8 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
+using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Forms;
+using static DwgCrawler.DwgParser;
+
 
 namespace DwgCrawler
 {
@@ -11,8 +17,12 @@ namespace DwgCrawler
 		[STAThread]
 		static void Main(string[] args)
 		{
-			var p =new DwgParser(args);
-			//p.WriteDwg("output.dwg");
+			var Args= Utils.ArgParser(args);
+			var p =new DwgParser( Args);
+			p.ExportJson();
+			
 		}
+		
+
 	}
 }
