@@ -1,6 +1,8 @@
 ﻿using System;
+using System.Diagnostics;
 using System.IO;
 using System.ServiceProcess;
+using System.Text;
 
 public class FileMonitorService
 {
@@ -12,7 +14,7 @@ public class FileMonitorService
 		{
 			Path = pathToWatch,
 			NotifyFilter = NotifyFilters.FileName | NotifyFilters.LastWrite,
-			Filter = "*.*",
+			Filter = "*.dwg",
 			EnableRaisingEvents = true
 		};
 
@@ -24,10 +26,46 @@ public class FileMonitorService
 	{
 		// Example action
 		File.AppendAllText(@"C:\Users\barako\Documents\file_events.log", $"{DateTime.Now}: {e.ChangeType} - {e.FullPath}\n");
+
 	}
 
 	public void StopMonitoring()
 	{
 		_watcher?.Dispose();
+	}
+	static async Task esAppend(string path)
+	{
+		// 1. Run the external executable and capture the JSON output
+		var startInfo = new ProcessStartInfo
+		{
+			FileName = "C:\\Users\\barako\\source\\repos\\ACadSharp\\src\\ACadSharp.Examples\\bin\\Debug\\net6.0-windows\\DwgCrawler.exe", // Replace with the actual executable path
+			Arguments = $"-j -f \"{path}\"", // Add arguments if needed
+			RedirectStandardOutput = true,
+			UseShellExecute = false,
+			CreateNoWindow = true
+		};
+
+		string jsonOutput;
+		using (var process = Process.Start(startInfo))
+		{
+			jsonOutput = await process.StandardOutput.ReadToEndAsync();
+			process.WaitForExit();
+		}
+
+		// 2. Send the JSON in a PUT request
+		using (var client = new HttpClient())
+		{
+			var content = new StringContent(jsonOutput, Encoding.UTF8, "application/json");
+
+			// Replace with your actual URL
+			string url = "https://your-api-endpoint.com/resource";
+
+			HttpResponseMessage response = await client.PutAsync(url, content);
+			string responseContent = await response.Content.ReadAsStringAsync();
+
+			Console.WriteLine($"Status: {response.StatusCode}");
+			Console.WriteLine("Response:");
+			Console.WriteLine(responseContent);
+		}
 	}
 }
